Move notifiable trait to event namespace
